-changed Weather obj due to incorrect data type for speed and degree of wind obj

+++ WeatherApp/Model/City.swift
@@ -66,5 +66,5 @@ struct Weather: Codable {
 
 // MARK: - Wind
 struct Wind: Codable {
-    let speed, deg: Int
+    let speed, deg: Double
 }
